+++ packages/modules/settings/src/models/index.ts
@@ -1,2 +1,3 @@
 export { ViewConfiguration } from "./view-configuration"
 export { UserPreference } from "./user-preference"
+export { PropertyLabel } from "./property-label"