+++ www/apps/api-reference/specs/admin/code_samples/Shell/admin_users_{id}_roles_{role_id}/delete.sh
@@ -1,2 +1,0 @@
-curl -X DELETE '{backend_url}/admin/users/{id}/roles/{role_id}' \
--H 'Authorization: Bearer {access_token}'